move search result item components to search page folder
