+++ apps/web/src/app/search/SearchResultItem/Image.tsx
