+++ apps/web/src/app/search/chat/RAGRetrievalResultItem/Image.tsx
