+++ apps/web/src/app/search/chat/RAGRetrievalResultItem/index.tsx
@@ -1,11 +1,11 @@
 import { matchRetrievalResult } from '@/Explorer/pattern'
 import { ExtractExplorerItem } from '@/Explorer/types'
 import { match } from 'ts-pattern'
-import AudioSearchItem from '../Search/Audio'
-import ImageSearchItem from '../Search/Image'
-import RawTextSearchItem from '../Search/RawText'
-import VideoSearchItem from '../Search/Video'
-import WebPageSearchItem from '../Search/WebPage'
+import AudioSearchItem from '../../SearchResultItem/Audio'
+import ImageSearchItem from '../../SearchResultItem/Image'
+import RawTextSearchItem from '../../SearchResultItem/RawText'
+import VideoSearchItem from '../../SearchResultItem/Video'
+import WebPageSearchItem from '../../SearchResultItem/WebPage'
 import AudioRetrievalItem from './Audio'
 import ImageRetrievalItem from './Image'
 import RawTextRetrievalItem from './RawText'
